Paper Sheet Completed

+++ MPC.Web/Areas/Api/Controllers/PaperSheetController.cs
@@ -1,5 +1,7 @@
 ﻿using System.Collections.Generic;
 using System.Linq;
+using System.Net;
+using System.Web;
 using System.Web.Http;
 using MPC.Interfaces.MISServices;
 using MPC.MIS.ModelMappers;
@@ -12,20 +14,34 @@ namespace MPC.MIS.Areas.Api.Controllers
 {
     public class PaperSheetController : ApiController
     {
+        #region Private
+
         private readonly IPaperSheetService paperSheetService;
 
+        #endregion
+
+        #region Constructor
+        /// <summary>
+        /// Constructor
+        /// </summary>
+        /// <param name="paperSheetService"></param>
         public PaperSheetController(IPaperSheetService paperSheetService)
         {
             this.paperSheetService = paperSheetService;
         }
+
+        #endregion
+
+        #region Public
+
         /// <summary>
         /// Get All Paper Sheets
         /// </summary>
         /// <returns></returns>
-        public PaperSheetResponseModel Get()
+        public PaperSheetResponseModel Get([FromUri] PaperSheetRequestModel request)
         {
             IEnumerable<PaperSize> paperSizes = null;
-            paperSizes = paperSheetService.GetAll();
+            paperSizes = paperSheetService.GetAll(request);
             var paperSheets = paperSizes as IList<PaperSize> ?? paperSizes.ToList();
             return new PaperSheetResponseModel
                    {
@@ -33,6 +49,7 @@ namespace MPC.MIS.Areas.Api.Controllers
                        RowCount = paperSheets.Count()
                    };
         }
+
         /// <summary>
         /// Update Paper Sheet
         /// </summary>
@@ -40,8 +57,13 @@ namespace MPC.MIS.Areas.Api.Controllers
         /// <returns></returns>
         public PaperSheet Put(PaperSheet paperSheet)
         {
-            return paperSheetService.Add(paperSheet.CreateFrom()).CreateFrom();
+            if (ModelState.IsValid)
+            {
+                return paperSheetService.Add(paperSheet.CreateFrom()).CreateFrom();
+            }
+            throw new HttpException((int) HttpStatusCode.BadRequest, "Invalid Request");
         }
+
         /// <summary>
         /// Create New Paper Sheet 
         /// </summary>
@@ -49,7 +71,11 @@ namespace MPC.MIS.Areas.Api.Controllers
         /// <returns></returns>
         public PaperSheet Post(PaperSheet paperSheet)
         {
-            return paperSheetService.Update(paperSheet.CreateFrom()).CreateFrom();
+            if (ModelState.IsValid)
+            {
+                return paperSheetService.Update(paperSheet.CreateFrom()).CreateFrom();
+            }
+            throw new HttpException((int) HttpStatusCode.BadRequest, "Invalid Request");
         }
 
         /// <summary>
@@ -59,7 +85,14 @@ namespace MPC.MIS.Areas.Api.Controllers
         /// <returns></returns>
         public bool Delete(PaperSheetRequestModel model)
         {
-            return paperSheetService.Delete(model.PaperSheetId);
+            if (ModelState.IsValid)
+            {
+                return paperSheetService.Delete(model.PaperSheetId);
+            }
+            throw new HttpException((int) HttpStatusCode.BadRequest, "Invalid Request");
         }
+
+        #endregion
+
     }
 }
